+++ cmd/prisimdemo/prisimdemo.go
@@ -7,6 +7,7 @@ package main
 import (
 	"fmt"
 	"image"
+	"log"
 	"os"
 
 	"image/draw"
@@ -31,6 +32,14 @@ func main() {
 	pf.RenderPanels()
 	draw.Draw(mockFrameBuffer, pf.Bounds, pf.Buffer, image.Point{}, draw.Src)
 	// Encode frame buffer as PNG and save
-	f, _ := os.Create("framebuffer.png")
-	png.Encode(f, mockFrameBuffer)
+	f, err := os.Create("framebuffer.png")
+	if err != nil {
+		log.Fatal(err)
+	}
+	if err := png.Encode(f, mockFrameBuffer); err != nil {
+		log.Fatal(err)
+	}
+	if err := f.Close(); err != nil {
+		log.Fatal(err)
+	}
 }